+++ task.py
@@ -23,6 +23,7 @@ from TASKdbcode import counttabledashboard
 from pretty_html_table import build_table
 import sqlalchemy
 import sys
+import textwrap
 from werkzeug.security import generate_password_hash,\
     check_password_hash
 
@@ -98,37 +99,45 @@ def selectmealsit1e():
     return response
  #-----------------------------------------------------------------------
 
-@app.route('/submitpatrondata', methods=['GET'])
+@app.route('/submitpatrondata', methods=['GET','POST'])
 @login_required(basic=True)
 def submitpatrondata():
-    # mealsite = request.args.get('mealsite')
     new_mealsite = request.args.get('mealsite')
-    mealsite = request.cookies.get('site')
-    set_new_mealsite = False         
-    # print("selected site")
-    # print(mealsite)
+    mealsite = request.cookies.get('mealsite')
+    new = False
+    num = request.cookies.get('num')
     if mealsite is None or (mealsite != new_mealsite and new_mealsite is not None): 
-        set_new_mealsite = True
+        new = True
         mealsite = new_mealsite
+        num = '0'
+    
+    submitted = request.form.get('language')
+    if submitted:
+        num = str(int(num)+1)   
+    
 
     races = []
-    # print(request.args.getlist('race'))
-    if request.args.getlist('race') is not None:
-        for race in request.args.getlist('race'):
-            print(race)
+    if request.form.getlist('race') is not None:
+        for race in request.form.getlist('race'):
             if (race != 'Unknown'):
-                races.append(race)
+                if race == "Native":
+                    races.append("Native Hawaiian/Pacific Islander")
+                elif race == "American":
+                    races.append("American Indian/Alaska Native")
+                else:
+                    races.append(race)
         races = list(filter(None, races))
+    print("RACE", races)
     racecsv = ",".join(races)
-    language = request.args.get('language')
-    age_range = request.args.get('age_range')
+    language = request.form.get('language')
+    age_range = request.form.get('age_range')
     # problem because the names changed
-    gender = request.args.get('gender')
-    zip_code = request.args.get('zip_codes')
-    homeless = request.args.get('homeless')
-    veteran = request.args.get('veteran')
-    disabled = request.args.get('disabled')
-    guessed = request.args.get('guessed')
+    gender = request.form.get('gender')
+    zip_code = request.form.get('zip_codes')
+    homeless = request.form.get('homeless')
+    veteran = request.form.get('veteran')
+    disabled = request.form.get('disabled')
+    guessed = request.form.get('guessed')
     # print('guess',guessed)
 
     patron_data = {"race": racecsv, "language": language,
@@ -136,9 +145,6 @@ def submitpatrondata():
     "homeless": homeless, "veteran": veteran, "disabled": disabled,
     "guessed": guessed}
 
-    # print(patron_data)
-
-    # print(any(patron_data.values()))
 
     zip_codes_by_mealsite = database_constants.ZIP_CODES[database_constants.MEAL_SITE_LOCATIONS[mealsite]]
     
@@ -159,13 +165,13 @@ def submitpatrondata():
         veteran_options = database_constants.VETERAN_OPTIONS,
         disabled_options = database_constants.DISABLED_OPTIONS,
         patron_response_options = database_constants.GUESSED_OPTIONS,
-        racecheck = ""
+        num = num,
         )
     response = make_response(html_code)
-
-    if set_new_mealsite:
-        response.set_cookie('site', mealsite)
-
+    response.set_cookie('num',num)
+    if new:
+        response.set_cookie('mealsite', new_mealsite)
+   
     return response
 
  #-----------------------------------------------------------------------
@@ -178,54 +184,71 @@ def admindisplaydata():
         "admin.html"
     )
 
-@app.route("/dash", methods=['POST','GET'])
-@login_required(must=[demographic_db.be_admin])
-def dash():
-    return app.index()
-
 
 
 @app.route('/register', methods=['GET', 'POST'])
 @login_required(must=[demographic_db.be_admin])
 def register(): 
-    email = request.args.get('email')
-    password = request.args.get('password')
-    account_type = request.args.get('AccountType')
-    repeat_password = request.args.get('repeatPassword')
-    if password != repeat_password: 
-        print("Passwords do not match")
-    account_details = {"username": email, "email": email, "password": password, "role": account_type}
-    demographic_db.add_user(account_details)
-
+    success = ''
+    username = request.form.get('username')
+    if username:
+        password = request.form.get('password')
+        account_type = 'representative'
+        repeat_password = request.form.get('repeatPassword')
+        if password != repeat_password: 
+            success = "Error: Passwords do not match"
+        else: 
+            account_details = {"username": username, "password": password, "role": account_type}
+            result = demographic_db.add_user(account_details)
+            if result:
+                success = "Success! Registered "+username+" as a TASK representative."
+            else:
+                success = "Error: Username taken, please try another."
     return render_template(
-        "register.html"
+        "register.html", success = success
     )
-
-@app.route('/viewusers', methods=['GET','POST'])
-@login_required(must=[demographic_db.be_admin])
-def viewusers(): 
-    return render_template("viewusers.html")
-
 
 @app.route('/users', methods=['GET','POST'])
 @login_required(must=[demographic_db.be_admin])
+def viewusers(): 
+    return render_template("userdashboard.html")
+
+@app.route('/viewusers', methods=['GET','POST'])
+@login_required(must=[demographic_db.be_admin])
 def users(): 
-    role = request.args.get('role')
-    df = demographic_db.get_users(role)
+    df = demographic_db.get_users()
     html = build_table(df, 'blue_light', padding='20px', even_color = 'black')
-    if role == 'administrators':
-        r = 'Administrator'
-    elif role == 'representatives':
-        r = 'Representative'
-    elif role == 'all':
-        r = 'All'
-    return render_template("users.html",table=html, titles=df.columns.values, role = r)
+    return render_template("viewusers.html",table=html, titles=df.columns.values, success = '')
+
+@app.route('/deleteusers', methods=['GET','POST'])
+@login_required(must=[demographic_db.be_admin])
+def deleteuser(): 
+    success = ''
+    username = request.form.get('user')
+    print(username)
+    if username:
+        if username == "jaimeparker":
+            success = 'You may not delete the administrator credentials from the system.'
+        else:
+            result = demographic_db.delete_user(username)
+            if not result:
+                success = 'User does not exist in system, please try again.'
+    df = demographic_db.get_users()
+    html = build_table(df, 'blue_light', padding='20px', even_color = 'black')
+    return render_template("deleteusers.html",table=html, titles=df.columns.values, role = 'All', success = success)
+
 
 @app.route('/deletelastpatron')
 @login_required(basic=True)
 def deletelast():
     meal_site = request.args.get('mealsite')
-    demographic_db.delete_last_patron(meal_site)
+    num = request.cookies.get('num')
+    if int(num) >0:
+        num = str(int(num) -1)
+        demographic_db.delete_last_patron(meal_site)
+    else:
+        num = '0'
+ 
     html_code = render_template('submitpatrondata.html',
         mealsite = meal_site,
         ampm=get_ampm(),
@@ -239,9 +262,10 @@ def deletelast():
         veteran_options = database_constants.VETERAN_OPTIONS,
         disabled_options = database_constants.DISABLED_OPTIONS,
         patron_response_options = database_constants.GUESSED_OPTIONS,
-        racecheck = ""
+        num = num
         )
     response = make_response(html_code)
+    response.set_cookie('num', num)
     return response
 
 @app.route('/getlastpatron')
@@ -249,8 +273,11 @@ def deletelast():
 def getlast():
     meal_site = request.args.get('mealsite')
     last = demographic_db.get_last_patron(meal_site)
+    lastrace = last['race'].iloc[0]
+    print(lastrace)
+    lastrace = "\n".join(textwrap.wrap(lastrace, width=20))
     html_code = render_template('prev.html',
-        lastrace = last['race'].iloc[0],
+        lastrace = lastrace,
         lastlanguage = last['language'].iloc[0],
         lastage = last['age_range'].iloc[0],
         lastgender = last['gender'].iloc[0],
